Add proto3 string enum support
In proto3, enums are canonically serialized as their string values.
While it's possible for libraries to allow this to be overridden to
use int values, that's not the default and is not particularly
useful for OpenAPI schema generation.

Source: https://protobuf.dev/programming-guides/proto3/#json-options

> **Emit enum values as integers instead of strings**
>
> The name of an enum value is used by default in JSON output.
> An option may be provided to use the numeric value of the enum
> value instead.

No support is added here for falling back to numeric values in
the generated schema. Similarly, the proto3 docs show an example
of attaching extra enum value options, such as a custom string
value. These could be added to `oapi/v1/field.proto` in the future
if needed.